requirements.txt
razbiutils==0.1.1
requests
pillow
python-dotenv
pystray
hidapi
pyee
Flask
Flask-Cors
waitress